refactor: move DocumentationUtils under Documentation folder
